feat: Simple Game Idea with three challenge levels
- Adds new launcher activity with challenge levels
- Replaces the hard-coded number of tiles & tries with flexible number based on challenge

The idea is to present the user with three challenge levels

1. Easy
2. Medium
3. Hard

Based on the user choice, set the number of tiles & max. number of tries. The user can play as many times as he wants in the current challenge level. To choose a much challenging level, he has to go back to the start of the app and change and continue playing. We can track how quickly user clears a challenge and award points/stars.

For now,

Easy -> 4 tiles, 4 tries
Medium -> 6 tiles, 6 tries
Hard -> 8 tiles, 8 tries

The number of tiles & tries are chosen as the same for simplicity but can be improved.