feat: Game with progressive levels
- Adds 'level' to the game to track the current level
- Shows 'current level' before starting a game
- Makes dialog dismiss listener as optional to avoid the need of additional listener
- Removes code imported from simple game branch

The idea of the game is to start with level 1 and increase difficulty as you win.

Level 1..3 -> 4 tiles & 4 tries
Level 4..7 -> 8 tiles & 8 tries
Level 8..20 -> 10 tiles & 10 tries

Games starts with level 1 and as the player wins the game, he's taken to next level. Once the max level is reached, the game will end. As of the max level is set to 10.